move convert methods
